Add files via upload
Hotel Vertical:
•	Every answer has its own row
o	When multiple values have been chosen for a multiple-choice question, each of the values will have its separate row.
•	More metadata that can be used like Location, Location Group, Question nr., Question-type, Question label, Response Id etc.
o	Response Id is a unique randomly generated combination of letters + numbers to identify a single respondent
o	I think the rest of the column headers are self-explanatory but don’t hesitate to ask when in doubt.
•	When someone stops in the middle of the questionnaire, all values till that moment are captured and extracted.

Hotel Horizontal:
•	All answers from one respondent are on one row.
•	Less metadata only Location, Location Group, Questions text, etc.
•	When someone stops in the middle of the questionnaire, all answers given will be available in the export, the rest of the cell’s will be empty.

Background information:
Budget hotel group with 20+ locations.
They have set a brand standard of 8.0 on several domains:
•	Reception (FrontDesk)
•	Room
•	Bathroom
These rating (1-10) question have conditional questions behind them, when a 1-6 or a 7-8 rating is given, you will receive an additional question to ask for any area’s which require improvement. These are separate questions for 1-6 and 7-8, this way we can create a clear distinction between area’s that need immediate attention and area’s that will suffice for the near future.

Apart from this, we usually work with the NPS Score, this is a value that can be calculated based on the NPS/Recommendation question which measures the loyalty of a consumer/guest.
The calculation is: %Promotors - % Detractors = NPS, the outcome is not a percentage, but simply the NPS value. Where:
•	Promoters (Green) = People who answered 9-10 on the NPS Question
•	Detractors (Red) = People who answered 0-6 on the NPS Question
•	Passives (Yellow) = People who answered 7-8 on the NPS Question but is not needed for the calculation.

There are some additional questions which give more information about the respondent instead of a rating or opinion of the hotel. These questions could be used to create a distinction between certain demographics such as people who stay there for business versus pleasure, etc….

Things that we would like to see:
•	Total amount of reviews, percentage of reviews from total per location
•	Average rating score per location with benchmark (benchmark = average score of all locations)
•	Average NPS score per location with benchmark (benchmark = average NPS of all locations)
•	Top 5 Locations / Worst 5 Locations - Average of Score and NPS
•	Top 5 Improvement points
•	% people for business & pleasure
•	% people for NPS types